// src/pages/DailyClinicalProcedurePlanningForm/types/DailyFields.ts
export type DailyFields = {
  data: string;
  id_medical_record: string;
  id_aluno_cd: string;
  id_aluno_asb: string;
  id_professor: string;
  avaliacao_professor?: string;
  aspectos_relevantes_da_anamnese: string;
  aspectos_relevantes_do_exame_radiografico: string;
  diagnostico: string;
  procedimento: string;
  protocolo_clinico: string;
};